feat: Update 'precio' column data type to INT in 'bebidas' and 'platos' tables
The 'precio' column in both the 'bebidas' and 'platos' tables was changed from DECIMAL(10, 2) to INT data type. This change was made to store the price values as integers instead of decimals, which simplifies the data structure and improves performance.

+++ src/main/resources/db/changelog/create_bebidas_table.sql
@@ -6,7 +6,7 @@ CREATE TABLE bebidas
     id            BIGINT                                   NOT NULL AUTO_INCREMENT PRIMARY KEY,
     nombre        VARCHAR(100)                             NOT NULL,
     descripcion   TEXT                                     NOT NULL,
-    precio        DECIMAL(10, 2)                           NOT NULL,
+    precio        INT                                      NOT NULL,
     stock         INT                                      NOT NULL DEFAULT 0,
     estado        enum ('ACTIVO', 'INACTIVO', 'ELIMINADO') NULL     DEFAULT 'ACTIVO',
     created_at    TIMESTAMP                                NULL     DEFAULT CURRENT_TIMESTAMP,
